Change classes.numcl to DECIMAL(2,1), charset to utf8mb4
Change number of classes (classes.numcl) column type to DECIMAL(2,1),
because for some 2022/2023 classes the number is 0.5 (!?).

+++ db/tables/classes.sql
@@ -1,13 +1,13 @@
 CREATE TABLE classes (
-  id        INT         NOT NULL,
-  name      VARCHAR(64) NOT NULL,
-  numcl     INT         NOT NULL COMMENT 'Number of classes',
-  yr        YEAR        NOT NULL COMMENT 'School year starting',
-  coef_bel  INT         NOT NULL COMMENT 'Coeficient for BEL result',
-  coef_mat  VARCHAR(45) NOT NULL COMMENT 'Coeficient for MAT result',
-  subj1_id  VARCHAR(5)  NOT NULL COMMENT 'Subject 1',
-  subj2_id  VARCHAR(5)  NOT NULL COMMENT 'Subject 2',
-  school_id INT         NOT NULL,
+  id        INT           NOT NULL,
+  name      VARCHAR(64)   NOT NULL,
+  numcl     DECIMAL(2,1)  NOT NULL COMMENT 'Number of classes',
+  yr        YEAR          NOT NULL COMMENT 'School year starting',
+  coef_bel  INT           NOT NULL COMMENT 'Coeficient for BEL result',
+  coef_mat  VARCHAR(45)   NOT NULL COMMENT 'Coeficient for MAT result',
+  subj1_id  VARCHAR(5)    NOT NULL COMMENT 'Subject 1',
+  subj2_id  VARCHAR(5)    NOT NULL COMMENT 'Subject 2',
+  school_id INT           NOT NULL,
 
   PRIMARY KEY (id, yr),
